Fixed MangaCatalog all extensions not working
