+++ src/ReadBerserk/ReadBerserk.ts
@@ -10,7 +10,6 @@ import {
 } from '../MangaCatalog'
 
 import { SourceBase } from '../MangaCatalogInterface'
-import { ReadBerserkParser } from './ReadBerserkParser'
 
 const DOMAIN = 'https://readberserk.com'
 
@@ -28,11 +27,6 @@ export const ReadBerserkInfo: SourceInfo = {
 }
 
 export class ReadBerserk extends MangaCatalog {
-
-    constructor(cheerio: CheerioAPI) {
-        super(cheerio)
-        this.parser = new ReadBerserkParser()
-    }
 
     baseUrl: string = DOMAIN
